requirements.txt
Django>=4.0.6,<4.1
djangorestframework>=3.13.1,<3.14
psycopg>=3.0.15,<3.1
drf-spectacular>=0.22.1,<0.23 
Pillow>=9.2.0,<9.3